eject script; enhanced readme

+++ packages/cca-scripts/scripts/destroy.js
@@ -1,3 +1,13 @@
+const inquirer = require('inquirer');
+
 const util = require('./util');
 
-util.cloudformation.deleteStack();
+inquirer.prompt({
+	type: 'confirm',
+	name: 'delete',
+	message: 'Are you sure you want to delete the API?'
+}).then(answer => {
+	if (answer.delete) {
+		util.cloudformation.deleteStack();
+	}
+});
